+++ activerecord/lib/active_record/relation.rb
@@ -5,8 +5,8 @@ module ActiveRecord
   # = Active Record Relation
   class Relation
     MULTI_VALUE_METHODS  = [:includes, :eager_load, :preload, :select, :group,
-                            :order, :joins, :where, :having, :bind, :references,
-                            :extending, :unscope]
+                            :order, :joins, :left_outer_joins, :where, :having, :bind,
+                            :references, :extending, :unscope]
 
     SINGLE_VALUE_METHODS = [:limit, :offset, :lock, :readonly, :from, :reordering,
                             :reverse_order, :distinct, :create_with, :uniq]